quiz_project/requirements.txt
django~=4.2.3
python-decouple
mysqlclient
drf-yasg
pytest~=7.4.0
djangorestframework~=3.14.0